Update Left Rotate the Array by One
Brute Force Approach: 
The extremely naive solution, we can think of, involves the use of an extra array. The algorithm is as follows.

Algorithm:
Suppose, there are N-X zeros and X non-zero elements in the array. We will first copy those X non-zero elements from the original array to a temporary array. 
Then, we will copy the elements from the temporary array one by one and fill the first X places of the original array. 
The last N-X places of the original array will be then filled with zero. Now, our task is completed.